health dyncfg split to separate file
